// libs/posts/src/lib/data/interfaces/post.interface.ts
import { Profile } from '@tt/data-access/profile'

export interface Post {
  id: number
  title: string
  communityId: number
  content: string
  author: Profile
  images: string[]
  createdAt: string
  updatedAt: string
  likes: number
  comments: PostComment[]
}

export interface PostCreateDto {
  title: string
  content: string
  authorId: number
  communityId: number
}

export interface PostComment {
  id: number
  text: string
  author: {
    id: number
    username: string
    avatarUrl: string
    subscribersAmount: number
  }
  postId: number
  commentId: number
  createdAt: string
  updatedAt: string
}

export interface CommentCreateDto {
  text: string
  authorId: number
  postId: number
  // commentId: number
}